+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Updated
 - `approximate_node_saturation`, `approximate_edge_saturation` and `approximate_saturation_curve` now uses the standard definition of saturation (s = 1 - molecules / reads).
 
+### Fixes
+- Fixed a bug in `SummarizeProximityScores` where the `include_missing_obs` argument was not properly handled when `include_missing_obs = FALSE`.
+
 ## [0.17.1] - 2026-04-10
 
 ### Updated